requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.1
discord.py==2.5.2
frozenlist==1.6.0
googletrans==4.0.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
python-dotenv
